W pliku tablica.py dopisano dtype. Dodano pliki macierz_zer_np_zeros.py oraz zwykla_petla_for_ma_slowko_range.py.

+++ tablica.py
@@ -1,6 +1,6 @@
 import numpy as np
 
 # Tworzenie tablicy z listy
-tablica = np.array([1, 2, 3, 4, 5])
+tablica = np.array([1, 2, 3, 4, 5], dtype = np.uint8)
 print(tablica)
 print(type(tablica)) # Zobaczysz <class 'numpy.ndarray'>
